clippy warnings and spelling fixes
